requirements.txt
flask==2.3.3
torch>=2.2.0
torchaudio>=2.2.0
torchvision>=0.17.0
transformers>=4.35.0
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.3
Pillow>=10.0.1
python-dotenv>=1.0.0
ftfy>=6.1.3
regex>=2023.8.8
tqdm>=4.66.1
matplotlib>=3.7.3
scipy>=1.11.3
requests>=2.31.0
pydub>=0.25.1
einops>=0.6.1
sentencepiece>=0.1.99
accelerate>=0.21.0
